refactor: fully separate OS kernel and componentize logic
This commit completes a deep architectural refactoring of the project.

- All core OS shell files (`App.tsx`, `types.ts`, `constants.tsx`,
  core UI components, contexts, and themes) are now encapsulated
  within the new `window/` (kernel) directory. The root directory
  is now clean, containing only project config and the main entry
  point.

- All import paths throughout the entire application have been
  updated to reflect this new structure.

- The `useWindowManager` hook now correctly handles all window
  management logic, abstracting it away from the main App component.

- The dynamic app discovery system is fully integrated, with apps
  defined in `.app` files and served by a backend API.
